Move options into separate mod
